2025-multi-agent-traffic/pyproject.toml
[project]
name = "marl-traffic-gen"
version = "0.1.0"
description = "Closed-loop Adversarial Multi-Agent Traffic Generation"
readme = "README.md"
authors = [
    { name = "Joshua Ransiek", email = "ransiek@fzi.de" }
]
requires-python = ">3.10,<3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "torch-summary>=1.4.5",
    "stable-baselines3>=2.3.2",
    "sb3-contrib>=2.3.0",
    "metadrive-simulator>=0.4.3",
    "torchinfo>=1.8.0",
    "tensorboard==2.11.2",
    "torch>=2.5.1",
    "rich>=13.9.2",
    "numpy<2.0.0",
    "pydantic>=2.11.1",
    "ray[rllib, tune]>=2.40.0",
    "ipywidgets>=8.1.7",
    "agent-trainers @ git+https://essgitlab.fzi.de/rl_testbench/2025-rl-agent-trainer.git",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "notebook>=7.2.2",
    "pdoc>=14.7.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "ruff>=0.6.5",
    "sphinx>=8.0.2",
    "tox-uv>=1.15.0",
    "tox>=4.19.0",
]

index-strategy = "unsafe-best-match"
extra-index-url = ["https://download.pytorch.org/whl/cu118/"]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118/"
explicit = true

[build-system]
requires = ["setuptools>=61.0", "wheel", "numpy>=1.22.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
    #"D",   # pydocstyle
    "A",   # builtins
    "RET", # returns
    "PTH", # pathlib instead of os
    "PL",  # pylint
    "PT",  # pytest-style
    "ANN",
]
ignore = [
    "D100",
    "A005", # Module shadows a Python standard-library module
    "F841", # Local variable is assigned to but never used
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["I001", "D104"]
"test*.py" = ["D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "google"
